Revert "refactor: 重构目录结果、github action流水线"
This reverts commit accd177a0fdf46048d98b624802dc61a15f3f6ac.
